requirements.txt
 azure-identity
 azure-mgmt-monitor
 azure-mgmt-loganalytics